Overrides default boto cache when run via lambda_handler
In lambda, the user home dir is read-only, so this is adding a mechanism
to override the default boto cache. When executed through the lambda handler,
the boto cache dir will be set to `/tmp/...`, which is writeable.

Also supports setting the boto cache dir through the BOTOCORE_CACHE_DIR env.